Added commands list file.
This file contains all commands in order to send periodically the input and output bytes from FastEthernet 0/0 and FastEthernet 1/0 interfaces.

The assumption about this commands list are the following:
-the router has FastEthernet 0/0 and FastEthernet 1/0 interfaces
-the FTFP server has instant-contiki username and 192.168.3.24 address

If the above assumptions doesn't match with the actual setting, they can be changed manually.

How does it work ? Copy and Paste and it works.